refactor(password): clean up imports and remove unused static constructors
- Remove static methods `fromPlain` and `fromHash` from Password value object
  - These methods were unused and added unnecessary abstraction
  - Direct constructor usage is now preferred for clarity

- Normalize import paths for StringValueObject
  - Ensure consistent casing and relative path resolution

- Simplify Password instantiation logic while preserving validation rules
